delete goods

+++ src/administrator.cpp
@@ -6,7 +6,10 @@
 
 #include <QSqlDatabase>
 #include <QSqlQuery>
-#include <QPointer>
+#include <QDate>
+#include <QTime>
+
+#include <QSqlError>
 
 Administrator::Administrator(QWidget *parent) :
     QWidget(parent),
@@ -77,8 +80,9 @@ void Administrator::on_goodsButton_clicked()
     while(query.next()) {
         int id = query.value(0).toInt();
         QString nama_barang = query.value(1).toString();
+        QString jumlah = query.value(2).toString();
         QString image_path = query.value(4).toString();
-        ListGoodsItem *item = new ListGoodsItem(id, QIcon(":/" + image_path), nama_barang);
+        ListGoodsItem *item = new ListGoodsItem(id, QIcon(":/" + image_path), nama_barang+ " " + jumlah);
         ui->listWidget_2->addItem(item);
         ++index;
     }
@@ -107,6 +111,8 @@ void Administrator::on_historyBack_clicked()
     ui->stackedWidget->setCurrentIndex(AdminWindow::NAVIGASI);
 }
 
+
+
 void Administrator::on_deleteItem_clicked()
 {
 
@@ -117,16 +123,67 @@ void Administrator::on_deleteItem_clicked()
         return;
     }
 
-    ListGoodsItem* current_item = dynamic_cast<ListGoodsItem* >(ui->listWidget_2->takeItem(current_row));
+    ListGoodsItem* current_item = dynamic_cast<ListGoodsItem* >(ui->listWidget_2->currentItem());
     int id = current_item->getId();
 
-    QSqlQuery query(QSqlDatabase::database("goods_connection"));
-    query.prepare("DELETE FROM goods WHERE id = :id");
-    query.bindValue(":id", id);
 
+    QSqlQuery query(QSqlDatabase::database("goods_connection"));
+    QSqlQuery history_query(QSqlDatabase::database("histories_connection"));
+
+    query.prepare("SELECT * FROM goods WHERE id = :id");
+    query.bindValue(":id", id);
     query.exec();
 
-    delete current_item;
+    int last_id, banyak_dihapus;
+    QString operasi, nama_barang, current_date, current_time;
+
+    history_query.exec("SELECT MAX(id) from history");
+    history_query.next();
+
+    last_id = history_query.value(0).toInt();
+    current_date = QDate::currentDate().toString("dd-MM-yyyy");
+    current_time = QTime::currentTime().toString();
+
+    if(query.next()) {
+        QString nama_barang = query.value(1).toString();
+        int jumlah = query.value(2).toInt();
+
+
+        if(jumlah == 1) {
+            banyak_dihapus =1;
+            operasi = "DELETE";
+
+            query.prepare("DELETE FROM goods WHERE id = :id ");
+            query.bindValue(":id", id);
+            query.exec();
+
+
+            ListGoodsItem* del = dynamic_cast<ListGoodsItem* >(ui->listWidget_2->takeItem(current_row));
+            delete del;
+        }
+        else {
+            jumlah = jumlah - 1;
+
+            query.prepare(QString("UPDATE goods SET jumlah = %1 WHERE id = :id").arg(QString::number(jumlah)));
+            query.bindValue(":id", id);
+            query.exec();
+            current_item->setText(nama_barang + " " + QString::number(jumlah));
+
+            banyak_dihapus = 1;
+            operasi = "DELETE";
+        }
+
+        history_query.prepare("INSERT INTO history ( id, operasi, nama_barang, jumlah, tanggal, waktu ) "
+                              "VALUES ( :id , :operasi , :nama_barang , :jumlah , :tanggal , :waktu )");
+        history_query.bindValue(":id", last_id+1);
+        history_query.bindValue(":operasi", "DELETE");
+        history_query.bindValue(":nama_barang", nama_barang);
+        history_query.bindValue(":jumlah", banyak_dihapus);
+        history_query.bindValue(":tanggal", current_date);
+        history_query.bindValue(":waktu", current_time);
+
+        history_query.exec();
+    }
 }
 
 void Administrator::on_deleteUser_clicked()
@@ -144,7 +201,7 @@ void Administrator::on_deleteUser_clicked()
 
     QSqlQuery query(QSqlDatabase::database("users_connection"));
 
-    query.prepare("DELETE FROM users WHERE id = :id");
+    query.prepare("SELECT * FROM users WHERE id = :id");
     query.bindValue(":id", id);
 
     query.exec();
